palette flashes have properly long interval in case of cl_maxfps is 999
git-svn-id: svn://localhost/repository@91 ad5b28d8-95b8-954d-b6c6-c0a63e036994

+++ trunk/client.h
@@ -49,7 +49,7 @@ typedef struct
 typedef struct
 {
 	int	destcolor[3];
-	int	percent;		// 0-256
+	double	percent;		// 0-256
 } cshift_t;
 
 #define	CSHIFT_CONTENTS	0
